many files: got rid of all warnings reported by -Wsign-compare abc-iw.c abc-wifi.c abc.c acache.c ad.c adht.c aip.c astart.c hash.c hash.h ai.c allnet_log.c allnet_queue.c app_util.c keys.c mapchar.c pipemsg.c pipemsg.h stream.c table.c trace_util.c util.c listen.c keyd.c sniffer.c traced.c record.c routing.c cutil.c retransmit.c store.c xchat_socket.c xcommon.c xcommon.h allnet_radio.c broadcast.c xtime.c

+++ src/abc/abc-iw.c
@@ -66,7 +66,8 @@ static int my_system (char * command)
     char * p = command;
     int found_blank = 0;
     argv [0] = command;
-    while (*p != '\0' && num_args <= sizeof (argv) / sizeof (char *)) {
+    while ((*p != '\0') &&
+           (num_args <= (int) (sizeof (argv) / sizeof (char *)))) {
       if (found_blank) {
         if (*p != ' ') {
           argv [num_args] = p;
@@ -79,7 +80,7 @@ static int my_system (char * command)
       }
       p++;
     }
-    if (num_args >= sizeof (argv) / sizeof (char *)) {
+    if (num_args >= (int) (sizeof (argv) / sizeof (char *))) {
       snprintf (alog->b, alog->s, "error: reading beyond argv %d\n", num_args);
       log_print (alog);
       argv [sizeof (argv) / sizeof (char *) - 1] = NULL;
